+++ src/multiset.h
@@ -14,7 +14,7 @@
 
 struct multiset
 {
-  INT32 refs;
+  ptrdiff_t refs;
 #ifdef PIKE_SECURITY
   struct object *prot;
 #endif
