+++ src/program.h
@@ -255,7 +255,7 @@ struct pike_trampoline
 
 struct program
 {
-  INT32 refs;
+  ptrdiff_t refs;
 #ifdef PIKE_SECURITY
   struct object *prot;
 #endif
